+++ src/test/java/com/cocodan/triplan/schedule/service/ScheduleMemberServiceTest.java
@@ -1,0 +1,24 @@
+package com.cocodan.triplan.schedule.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
+
+@SpringBootTest
+@Transactional
+public class ScheduleMemberServiceTest {
+
+    @Autowired
+    private ScheduleMemberService scheduleMemberService;
+
+    @Autowired
+    private ScheduleService scheduleService;
+
+    //여행멤버 추가하기
+
+    // 여행멤버 리스트 조회하기
+
+    // 여행멤버 삭제하기
+
+    // 여행에서 나가기
+}
